Move ConfigurationProviderDatabase to Yousei.Yousei.Internal.Database
